Extracted Predefined*** classes into a separate module.
